+++ RedPandaIDE/editor.cpp
@@ -114,7 +114,7 @@ Editor::Editor(QWidget *parent, const QString& filename,
     mFileEncoding = ENCODING_ASCII;
     if (!isNew) {
         try {
-            loadFile();
+            loadContent();
         } catch (FileError& e) {
             QMessageBox::critical(nullptr,
                                   tr("Error Load File"),
@@ -225,9 +225,11 @@ Editor::Editor(QWidget *parent, const QString& filename,
     //show event is trigged when this is added to the qtabwidget
     if (!pMainWindow->openingFiles()
             && !pMainWindow->openingProject()) {
-        reparse(false);
-        checkSyntaxInBack();
-        reparseTodo();
+        if (inProject()) {
+            reparse(false);
+            reparseTodo();
+        }
+        //checkSyntaxInBack();
     }
 }
 
@@ -258,25 +260,15 @@ void Editor::loadFile(QString filename) {
 //        }
     } else {
         filename = QFileInfo(filename).absoluteFilePath();
+        doSetFileType(getFileType(filename));
     }
 
     //FileError should by catched by the caller of loadFile();
-
-    this->document()->loadFromFile(filename,mEncodingOption,mFileEncoding);
-
-    if (mProject) {
-        PProjectUnit unit = mProject->findUnit(this);
-        if (unit) {
-            unit->setRealEncoding(mFileEncoding);
-        }
-    }
-
-    //this->setModified(false);
+    mFilename = filename;
+    loadContent();
     updateCaption();
-    if (inTab())
-        pMainWindow->updateForEncodingInfo(this);
-    doSetFileType(getFileType(mFilename));
-    applyColorScheme(pSettings->editor().colorScheme());
+
+//    applyColorScheme(pSettings->editor().colorScheme());
     if (!inProject()) {
         initParser();
         reparse(false);
@@ -285,8 +277,6 @@ void Editor::loadFile(QString filename) {
     if (pSettings->editor().syntaxCheckWhenLineChanged()) {
         checkSyntaxInBack();
     }
-
-    saveAutoBackup();
 }
 
 void Editor::saveFile(QString filename) {
@@ -543,6 +533,8 @@ const QByteArray& Editor::encodingOption() const noexcept{
     return mEncodingOption;
 }
 void Editor::setEncodingOption(const QByteArray& encoding) noexcept{
+    if (encoding.isEmpty())
+        return;
     QByteArray newEncoding=encoding;
     if (mProject && encoding==ENCODING_PROJECT)
         newEncoding=mProject->options().encoding;
@@ -1505,23 +1497,23 @@ void Editor::closeEvent(QCloseEvent *)
 
 void Editor::showEvent(QShowEvent */*event*/)
 {
-    if (mParser && pSettings->codeCompletion().clearWhenEditorHidden()
-            && pSettings->codeCompletion().shareParser()
-            && !inProject()
-            && isC_CPPSourceFile(mFileType) ) {
-        //qDebug()<<mFilename;
-        resetCppParser(mParser);
-    }
     if (mParser && !pMainWindow->isClosingAll()
-            && !pMainWindow->isQuitting()
-            && !mParser->isFileParsed(mFilename)
-            ) {
+            && !pMainWindow->isQuitting()) {
         connect(mParser.get(),
                 &CppParser::onEndParsing,
                 this,
                 &Editor::onEndParsing);
-        if (!pMainWindow->openingFiles() && !pMainWindow->openingProject())
-            reparse(false);
+        if (!pMainWindow->openingFiles() && !pMainWindow->openingProject()) {
+            if (pSettings->codeCompletion().clearWhenEditorHidden()
+                && pSettings->codeCompletion().shareParser()
+                && !inProject()) {
+                if (needReparse())
+                    resetCppParser(mParser);
+            }
+            if (needReparse()) {
+                reparse(false);
+            }
+        }
     }
     if (inTab()) {
         pMainWindow->debugger()->setIsForProject(inProject());
@@ -2163,6 +2155,24 @@ void Editor::onEndParsing()
     mIdentCache.clear();
     document()->invalidateAllNonTempLineWidth();
     invalidate();
+}
+
+void Editor::loadContent()
+{
+    this->document()->loadFromFile(mFilename,mEncodingOption,mFileEncoding);
+    applyColorScheme(pSettings->editor().colorScheme());
+    if (mProject) {
+        PProjectUnit unit = mProject->findUnit(this);
+        if (unit) {
+            unit->setRealEncoding(mFileEncoding);
+        }
+    }
+    //this->setModified(false);
+
+    if (inTab())
+        pMainWindow->updateForEncodingInfo(this);
+
+    saveAutoBackup();
 }
 
 void Editor::resolveAutoDetectEncodingOption()
@@ -3090,7 +3100,7 @@ void Editor::initParser()
         if (pSettings->codeCompletion().shareParser()) {
             mParser = sharedParser(calcParserLanguage());
         } else {
-            mParser = CppParser::createParser();
+            mParser = std::make_shared<CppParser>();
             mParser->setLanguage(calcParserLanguage());
             mParser->setOnGetFileStream(
                         std::bind(
@@ -4537,18 +4547,18 @@ void Editor::setFileType(FileType newFileType)
     if (mFileType==newFileType)
         return;
     doSetFileType(newFileType);
-    applyColorScheme(pSettings->editor().colorScheme());
     if (!inProject()) {
         initParser();
         reparse(false);
     }
     reparseTodo();
     updateCaption();
+    pMainWindow->updateEditorActions(this);
 }
 
-void Editor::doSetFileType(FileType newFileType)
+void Editor::doSetFileType(FileType newFileType, bool force)
 {
-    if (mFileType==newFileType)
+    if (mFileType==newFileType && !force)
         return;
     mFileType = newFileType;
 
@@ -4558,6 +4568,17 @@ void Editor::doSetFileType(FileType newFileType)
         break;
     case FileType::CSource:
         mUseCppSyntax = false;
+        break;
+    case FileType::CCppHeader:
+    {
+        FileType contextFileType = getFileType(mContextFile);
+        if (contextFileType==FileType::CppSource)
+            mUseCppSyntax = true;
+        else if (contextFileType == FileType::CSource)
+            mUseCppSyntax = false;
+        else
+            mUseCppSyntax = pSettings->editor().defaultFileCpp();
+    }
         break;
     default:
         mUseCppSyntax = pSettings->editor().defaultFileCpp();
@@ -4573,6 +4594,7 @@ void Editor::doSetFileType(FileType newFileType)
     }
 
     setSyntaxer(syntaxer);
+    applyColorScheme(pSettings->editor().colorScheme());
 }
 
 Editor* Editor::openFileInContext(const QString &filename)
@@ -4589,8 +4611,18 @@ Editor* Editor::openFileInContext(const QString &filename)
                 e->setContextFile(mContextFile);
             }
         }
+        if (e->isVisible())
+            pMainWindow->updateClassBrowserForEditor(e);
     }
     return e;
+}
+
+bool Editor::needReparse()
+{
+    return mParser && ((isC_CPPHeaderFile(mFileType)
+                        && !mContextFile.isEmpty()
+                        && !mParser->isFileParsed(mContextFile))
+                       || (!mParser->isFileParsed(mFilename)));
 }
 
 const QString &Editor::contextFile() const
@@ -4600,13 +4632,13 @@ const QString &Editor::contextFile() const
 
 void Editor::setContextFile(const QString &newContextFile)
 {
-    if (!isC_CPPHeaderFile(mFileType))
-        return;
     QString s = newContextFile.trimmed();
     if (mContextFile == s)
         return;
     mContextFile = s;
-    if (!mContextFile.isEmpty()) {
+    if (isC_CPPHeaderFile(mFileType)) {
+        doSetFileType(mFileType, true);
+        applyColorScheme(pSettings->editor().colorScheme());
         reparse(false);
     }
 }
@@ -4623,7 +4655,7 @@ PCppParser Editor::sharedParser(ParserLanguage language)
         parser=mSharedParsers[language].lock();
     }
     if (!parser) {
-        parser=CppParser::createParser();
+        parser = std::make_shared<CppParser>();
         parser->setLanguage(language);
         parser->setOnGetFileStream(
                     std::bind(